forecast-prophet_docker/requirements.txt
pandas
python-dotenv
prophet
plotly>=5.0.0
kaleido